requirements.txt
## for traditional
Pillow
scipy
numpy


## for photomosaic
## These are required for developing the package (running the tests, building
## the documentation) but not necessarily required for _using_ it.
# codecov
# coverage
# flake8
# pytest >3.9
# sphinx
## These are dependencies of various sphinx extensions for documentation.
# ipython
# matplotlib
# numpydoc
# sphinx-copybutton
# sphinx_rtd_theme
## requirements
colorspacious
#numpy
requests
scikit-image
#scipy
tqdm
photomosaic